requirements.txt
anyio==4.12.1
certifi==2026.1.4
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.4.1
openpyxl==3.1.5
pandas==2.3.3
python-dateutil==2.9.0.post0
python-telegram-bot==22.5
pytz==2025.2
six==1.17.0
tzdata==2025.3
